rotas funcionando

+++ src/main/java/com/celliaraujo/api_cadastro_produtos/service/ProdutoService.java
@@ -2,7 +2,9 @@ package com.celliaraujo.api_cadastro_produtos.service;
 
 import com.celliaraujo.api_cadastro_produtos.model.Produto;
 import com.celliaraujo.api_cadastro_produtos.repository.ProdutoRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,7 +30,14 @@ public class ProdutoService{
     }
 
     public void excluirProduto(Long id){
+        if(id == null){
+            throw new IllegalArgumentException("O id não pode ser nulo.");
+        }
+        if (!produtoRepository.existsById(id)) {
+            throw new EntityNotFoundException("Produto com id " + id + " não encontrado.");
+        }
         produtoRepository.deleteById(id);
+
     }
